check if volume control enabled on mac

+++ src/term.c
@@ -39,7 +39,7 @@ void setTextColor(int color)
 
 void setTextColorRGB(int r, int g, int b)
 {
-        printf("\033[0;38;2;%03u;%03u;%03um", r, g, b);
+        printf("\033[0;38;2;%03u;%03u;%03um", (unsigned int)r, (unsigned int)g, (unsigned int)b);
 }
 
 void getTermSize(int *width, int *height)
